Move AppConfig into app-commons, reduce classpath deps on keyManager,tor
