+++ app-commons/src/main/scala/org/bitcoins/commons/util/DatadirUtil.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.db.util
+package org.bitcoins.commons.util
 
 import com.typesafe.config.Config
 import org.bitcoins.core.config._
