+++ app/bundle/src/main/scala/org/bitcoins/bundle/gui/LandingPane.scala
@@ -2,7 +2,7 @@ package org.bitcoins.bundle.gui
 
 import akka.actor.ActorSystem
 import grizzled.slf4j.Logging
-import org.bitcoins.db.util.ServerArgParser
+import org.bitcoins.commons.util.ServerArgParser
 import org.bitcoins.gui._
 import org.bitcoins.node.NodeType
 import org.bitcoins.server.BitcoinSAppConfig
