+++ app/server-routes/src/main/scala/org/bitcoins/server/routes/BitcoinSRunner.scala
@@ -2,8 +2,8 @@ package org.bitcoins.server.routes
 
 import akka.actor.ActorSystem
 import grizzled.slf4j.Logging
+import org.bitcoins.commons.util.ServerArgParser
 import org.bitcoins.core.util.{EnvUtil, StartStopAsync}
-import org.bitcoins.db.util.ServerArgParser
 
 import scala.concurrent.{ExecutionContext, Future}
