+++ app/server-routes/src/main/scala/org/bitcoins/server/routes/Server.scala
@@ -8,7 +8,7 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server._
 import akka.http.scaladsl.server.directives.DebuggingDirectives
 import de.heikoseeberger.akkahttpupickle.UpickleSupport._
-import org.bitcoins.db.AppConfig
+import org.bitcoins.commons.config.AppConfig
 import upickle.{default => up}
 
 import scala.concurrent.Future
