+++ db-commons/src/main/scala/org/bitcoins/db/DbAppConfig.scala
@@ -1,7 +1,7 @@
 package org.bitcoins.db
 
 import com.typesafe.config._
-import org.bitcoins.db.AppConfig.safePathToString
+import org.bitcoins.commons.config._
 import org.bitcoins.db.DatabaseDriver.{PostgreSQL, SQLite}
 import slick.basic.DatabaseConfig
 import slick.jdbc.JdbcProfile
@@ -34,7 +34,7 @@ abstract class DbAppConfig extends AppConfig {
   lazy val jdbcUrl: String = {
     driver match {
       case SQLite =>
-        s""""jdbc:sqlite:"${safePathToString(dbPath)}/$dbName"""
+        s""""jdbc:sqlite:"${AppConfig.safePathToString(dbPath)}/$dbName"""
       case PostgreSQL =>
         s""""jdbc:postgresql://$dbHost:$dbPort/$dbName""""
     }
@@ -96,7 +96,7 @@ abstract class DbAppConfig extends AppConfig {
          |bitcoin-s {
          |  $moduleName {
          |     db {
-         |        path = ${safePathToString(dbPath)}
+         |        path = ${AppConfig.safePathToString(dbPath)}
          |        name = $dbName
          |        url = $jdbcUrl
          |     }
