+++ testkit/src/main/scala/org/bitcoins/testkit/wallet/BitcoinSDualWalletTest.scala
@@ -1,8 +1,8 @@
 package org.bitcoins.testkit.wallet
 
+import org.bitcoins.commons.config.AppConfig
 import org.bitcoins.core.currency.Satoshis
 import org.bitcoins.core.protocol.dlc.models.{ContractInfo, ContractOraclePair}
-import org.bitcoins.db.AppConfig
 import org.bitcoins.dlc.wallet.DLCAppConfig
 import org.bitcoins.server.BitcoinSAppConfig
 import org.bitcoins.testkit.BitcoinSTestAppConfig
