+++ tor/src/main/scala/org/bitcoins/tor/config/TorAppConfig.scala
@@ -1,8 +1,8 @@
 package org.bitcoins.tor.config
 
 import com.typesafe.config.Config
+import org.bitcoins.commons.config.{AppConfig, AppConfigFactory, ConfigOps}
 import org.bitcoins.core.util.NetworkUtil
-import org.bitcoins.db.{AppConfig, AppConfigFactory, ConfigOps}
 import org.bitcoins.tor.TorProtocolHandler.{Password, SafeCookie}
 import org.bitcoins.tor.{Socks5ProxyParams, TorParams}
